compute move template
